Refactor Header.tsx
