+++ src/Pages/Authenticated/Sessions/Sessions.tsx
@@ -16,7 +16,7 @@ import { parseError } from '../../../commons/error.utils'
 import formatBytes from '../../../commons/formatBytes'
 import { displayMyst } from '../../../commons/money.utils'
 import { toastError } from '../../../commons/toast.utils'
-import Header from '../../../Components/Header'
+import Header from '../../../Components/Header/Header'
 import { RootState } from '../../../redux/store'
 import SessionSidebar from '../SessionSidebar/SessionSidebar'
 import './Sessions.scss'
